Split UI-related into a separate file.
Some of these are a bit bigger than just UI libraries, but this division
seems close enough.
